Crate darwin_rs

darwin-rs: evolutionary algorithms with Rust

Written by Willi Kappler, Version 0.2 (2016.08.xx)

Repository: https://github.com/willi-kappler/darwin-rs

License: MIT

This library allows you to write evolutionary algorithms (EA) in Rust. Examples provided: TSP, Sudoku, Queens Problem

Reexports

pub use individual::Individual;

pub use simulation::Simulation;

pub use simulation_builder::{SimulationBuilder, SimError};

pub use population::Population;

pub use population_builder::{PopulationBuilder, PopError};

Modules

individual
This module defines the trait and the data structure needed for specifying each individual in a population.

population
This module defines structure and methods for a population that is needed by a smulation.

population_builder
This module defines helper functions (builder pattern) to create a valid population.

simulation
This module defines structures and methods for an EA simulation.

simulation_builder
This module defines helper functions (builder pattern) to create a valid simulation.