Enum darwin_rs::BuilderResult

pub enum BuilderResult<T: Individual + Send + Sync> {
    TooLowEndIterration,
    TooLowIndividuals,
    Ok(Simulation<T>),
}

This enum describes the possible return error that the simulation builder can return.

Variants

TooLowEndIterration

The number of iteration is too low, should be >= 10

TooLowIndividuals

The number of individuals is too loe, should be >= 3

Ok

Everything is fine, the simulation is properly configured and ready to run.