Struct darwin_rs::individual::IndividualWrapper

pub struct IndividualWrapper<T: Individual> {
    pub individual: T,
    pub fitness: f64,
    pub num_of_mutations: u32,
    pub id: u32,
}

A wrapper helper struct for the individuals. It does the book keeping of the fitness and the number of mutations this individual has to run in one iteration.

Fields

individual

The actual individual, user defined struct.

fitness

The current calculated fitness for this individual.

num_of_mutations

The number of mutation this individual is doing in one iteration.

id

The id of the population that this individual belongs to. Just for statistics