Struct darwin_rs::simulation::SimulationResult

pub struct SimulationResult<T: Individual + Send + Sync> {
    pub improvement_factor: f64,
    pub original_fitness: f64,
    pub fittest: Vec<IndividualWrapper<T>>,
    pub iteration_counter: u32,
}

The SimulationResult Type. Holds the simulation results: All the fittest individuals, the improvement_factor, the iteration_counter and the original_fitness.

Fields

improvement_factor

The current improvement factor, that means the ration between the very first and the current fitness.

original_fitness

The very first calculated fitness, when the simulation just started.

fittest

Vector of fittest individuals. This will change during the simulation as soon as a new more fittest individual is found and pushed into the first position (index 0).

iteration_counter

How many iteration did the simulation run